# Asset-Manager/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Beacon Nonprofit Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /supporters:
    get:
      operationId: listSupporters
      tags: [supporters]
      summary: List supporters (donors)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: supportType
          in: query
          schema:
            type: string
        - name: segment
          in: query
          schema:
            type: string
        - name: churnRisk
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Supporter list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupporterListResponse"
    post:
      operationId: createSupporter
      tags: [supporters]
      summary: Create supporter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupporterBody"
      responses:
        "201":
          description: Created supporter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supporter"
  /supporters/{id}:
    get:
      operationId: getSupporter
      tags: [supporters]
      summary: Get supporter by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supporter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supporter"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSupporter
      tags: [supporters]
      summary: Update supporter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupporterBody"
      responses:
        "200":
          description: Updated supporter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supporter"
    delete:
      operationId: deleteSupporter
      tags: [supporters]
      summary: Delete supporter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /supporters/{id}/giving-stats:
    get:
      operationId: getSupporterGivingStats
      tags: [supporters]
      summary: Get giving statistics for a supporter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Giving statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GivingStats"
...
  /social-media-posts:
    get:
      operationId: listSocialMediaPosts
      tags: [socialMedia]
      summary: List social media posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: platform
          in: query
          schema:
            type: string
        - name: postType
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Social media post list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialMediaPostListResponse"
    post:
      operationId: createSocialMediaPost
      tags: [socialMedia]
      summary: Create social media post record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSocialMediaPostBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialMediaPost"
  /social-media-posts/{id}:
    get:
      operationId: getSocialMediaPost
      tags: [socialMedia]
      summary: Get social media post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialMediaPost"
    patch:
      operationId: updateSocialMediaPost
      tags: [socialMedia]
      summary: Update social media post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSocialMediaPostBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialMediaPost"
    delete:
      operationId: deleteSocialMediaPost
      tags: [socialMedia]
      summary: Delete social media post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /social-media-posts/analytics:
    get:
      operationId: getSocialMediaAnalytics
      tags: [socialMedia]
      summary: Social media analytics summary
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialMediaAnalytics"

  # ─── IMPACT SNAPSHOTS ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]

    # ─── AUTH ───────────────────────────────────────────────────────────────
...
    Supporter:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        organization:
          type: ["string", "null"]
        supportType:
          type: string
          enum: [individual, corporate, foundation, government, other]
        acquisitionChannel:
          type: ["string", "null"]
        segment:
          type: ["string", "null"]
        churnRiskScore:
          type: ["number", "null"]
        upgradeScore:
          type: ["number", "null"]
        lifetimeGiving:
          type: number
        lastGiftDate:
          type: ["string", "null"]
        lastGiftAmount:
          type: ["number", "null"]
        isRecurring:
          type: boolean
        communicationPreference:
          type: ["string", "null"]
        interests:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, firstName, lastName, email, supportType, lifetimeGiving, isRecurring]
    CreateSupporterBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        organization:
          type: string
        supportType:
          type: string
          enum: [individual, corporate, foundation, government, other]
        acquisitionChannel:
          type: string
        communicationPreference:
          type: string
        interests:
          type: array
          items:
            type: string
      required: [firstName, lastName, email, supportType]
    UpdateSupporterBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        organization:
          type: string
        supportType:
          type: string
        acquisitionChannel:
          type: string
        communicationPreference:
          type: string
        interests:
          type: array
          items:
            type: string
    SupporterListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Supporter"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
    GivingStats:
      type: object
      properties:
        supporterId:
          type: integer
        lifetimeGiving:
          type: number
        givingThisYear:
          type: number
        lastGiftDate:
          type: ["string", "null"]
        lastGiftAmount:
          type: ["number", "null"]
        numberOfGifts:
          type: integer
        averageGiftSize:
          type: number
        largestGift:
          type: number
        campaignsSupported:
          type: integer
        isRecurring:
          type: boolean
        monthlyTrend:
          type: array
          items:
            $ref: "#/components/schemas/DonationTrend"
      required: [supporterId, lifetimeGiving, givingThisYear, numberOfGifts, averageGiftSize, largestGift, isRecurring]
...
    DonationTrend:
      type: object
      properties:
        month:
          type: string
        totalAmount:
          type: number
        donationCount:
          type: integer
        avgAmount:
          type: number
      required: [month, totalAmount, donationCount, avgAmount]
...
    SocialMediaPost:
      type: object
      properties:
        id:
          type: integer
        platform:
          type: string
          enum: [facebook, instagram, twitter, linkedin, tiktok]
        postType:
          type: string
        content:
          type: string
        postDate:
          type: string
        timeWindow:
          type: ["string", "null"]
        likes:
          type: integer
        shares:
          type: integer
        comments:
          type: integer
        reach:
          type: integer
        engagementRate:
          type: number
        donationReferrals:
          type: integer
        donationValueFromPost:
          type: number
        predictedConversionScore:
          type: ["number", "null"]
        createdAt:
          type: string
      required: [id, platform, postType, content, postDate, likes, shares, comments, reach, engagementRate, donationReferrals, donationValueFromPost]
    CreateSocialMediaPostBody:
      type: object
      properties:
        platform:
          type: string
          enum: [facebook, instagram, twitter, linkedin, tiktok]
        postType:
          type: string
        content:
          type: string
        postDate:
          type: string
        timeWindow:
          type: string
        likes:
          type: integer
        shares:
          type: integer
        comments:
          type: integer
        reach:
          type: integer
        donationReferrals:
          type: integer
        donationValueFromPost:
          type: number
      required: [platform, postType, content, postDate]
    UpdateSocialMediaPostBody:
      type: object
      properties:
        platform:
          type: string
        postType:
          type: string
        content:
          type: string
        postDate:
          type: string
        timeWindow:
          type: string
        likes:
          type: integer
        shares:
          type: integer
        comments:
          type: integer
        reach:
          type: integer
        donationReferrals:
          type: integer
        donationValueFromPost:
          type: number
    SocialMediaPostListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SocialMediaPost"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
    SocialMediaAnalytics:
      type: object
      properties:
        postsThisMonth:
          type: integer
        avgEngagementRate:
          type: number
        donationReferrals:
          type: integer
        donationValueFromSocial:
          type: number
        bestPlatform:
          type: string
        bestPostType:
          type: string
        bestTimeWindow:
          type: string
        platformBreakdown:
          type: array
          items:
            type: object
            properties:
              platform:
                type: string
              posts:
                type: integer
              avgEngagement:
                type: number
              donationReferrals:
                type: integer
            required: [platform, posts, avgEngagement, donationReferrals]
        engagementHeatmap:
          type: array
          items:
            type: object
            properties:
              dayOfWeek:
                type: integer
              hour:
                type: integer
              avgEngagement:
                type: number
            required: [dayOfWeek, hour, avgEngagement]
      required: [postsThisMonth, avgEngagementRate, donationReferrals, donationValueFromSocial, bestPlatform, bestPostType, bestTimeWindow]

    # ─── IMPACT SNAPSHOT ─────────────────────────────────────────────────────
...